polishing and refactors
