+++ src/readiness/constants.ts
@@ -1,0 +1,12 @@
+// Splits events:
+export const SDK_SPLITS_ARRIVED = 'SDK_SPLITS_ARRIVED';
+export const SDK_SPLITS_CACHE_LOADED = 'SDK_SPLITS_CACHE_LOADED';
+
+// Segments events:
+export const SDK_SEGMENTS_ARRIVED = 'SDK_SEGMENTS_ARRIVED';
+
+// Readiness events:
+export const SDK_READY_TIMED_OUT = 'init::timeout';
+export const SDK_READY = 'init::ready';
+export const SDK_READY_FROM_CACHE = 'init::cache-ready';
+export const SDK_UPDATE = 'state::update';
