+++ src/readiness/readinessManager.ts
@@ -1,19 +1,7 @@
 import objectAssign from 'object-assign';
 import { IEventEmitter } from '../types';
+import { SDK_SPLITS_ARRIVED, SDK_SPLITS_CACHE_LOADED, SDK_SEGMENTS_ARRIVED, SDK_READY_TIMED_OUT, SDK_READY_FROM_CACHE, SDK_UPDATE, SDK_READY } from './constants';
 import { IReadinessEventEmitter, IReadinessManager, ISegmentsEventEmitter, ISplitsEventEmitter } from './types';
-
-// Splits events:
-export const SDK_SPLITS_ARRIVED = 'SDK_SPLITS_ARRIVED';
-export const SDK_SPLITS_CACHE_LOADED = 'SDK_SPLITS_CACHE_LOADED';
-
-// Segments events:
-export const SDK_SEGMENTS_ARRIVED = 'SDK_SEGMENTS_ARRIVED';
-
-// Readiness events:
-export const SDK_READY_TIMED_OUT = 'init::timeout';
-export const SDK_READY = 'init::ready';
-export const SDK_READY_FROM_CACHE = 'init::cache-ready';
-export const SDK_UPDATE = 'state::update';
 
 function splitsEventEmitterFactory(EventEmitter: new () => IEventEmitter): ISplitsEventEmitter {
   const splitsEventEmitter = objectAssign(new EventEmitter(), {
@@ -21,7 +9,10 @@ function splitsEventEmitterFactory(EventEmitter: new () => IEventEmitter): ISpli
     splitsCacheLoaded: false,
   });
 
-  splitsEventEmitter.once(SDK_SPLITS_ARRIVED, () => { splitsEventEmitter.splitsArrived = true; });
+  // `isSplitKill` condition avoids an edge-case of wrongly emitting SDK_READY if:
+  // - `/mySegments` fetch and SPLIT_KILL occurs before `/splitChanges` fetch, and
+  // - storage has cached splits (for which case `splitsStorage.killLocally` can return true)
+  splitsEventEmitter.once(SDK_SPLITS_ARRIVED, (isSplitKill) => { if (!isSplitKill) splitsEventEmitter.splitsArrived = true; });
   splitsEventEmitter.once(SDK_SPLITS_CACHE_LOADED, () => { splitsEventEmitter.splitsCacheLoaded = true; });
 
   return splitsEventEmitter;
