+++ src/readiness/sdkReadinessManager.ts
@@ -1,15 +1,10 @@
 import objectAssign from 'object-assign';
 import promiseWrapper from '../utils/promise/wrapper';
-import {
-  readinessManagerFactory,
-  SDK_READY,
-  SDK_READY_FROM_CACHE,
-  SDK_UPDATE,
-  SDK_READY_TIMED_OUT
-} from './readinessManager';
+import { readinessManagerFactory } from './readinessManager';
 import { ISdkReadinessManager } from './types';
 import { IEventEmitter } from '../types';
 import { logFactory } from '../logger/sdkLogger';
+import { SDK_READY, SDK_READY_TIMED_OUT, SDK_READY_FROM_CACHE, SDK_UPDATE } from './constants';
 const log = logFactory('');
 
 const NEW_LISTENER_EVENT = 'newListener';
