+++ src/sync/offline/syncTasks/fromObjectSyncTask.ts
@@ -8,6 +8,7 @@ import syncTaskFactory from '../../syncTask';
 import { ISyncTask } from '../../types';
 import { ISettings } from '../../../types';
 import { CONTROL } from '../../../utils/constants';
+import { SDK_SPLITS_ARRIVED, SDK_SEGMENTS_ARRIVED } from '../../../readiness/constants';
 const log = logFactory('splitio-producer:offline');
 
 /**
@@ -55,8 +56,8 @@ export function fromObjectUpdaterFactory(
         storage.splits.clear(),
         storage.splits.addSplits(splits)
       ]).then(() => {
-        readiness.splits.emit('SDK_SPLITS_ARRIVED');
-        readiness.segments.emit('SDK_SEGMENTS_ARRIVED');
+        readiness.splits.emit(SDK_SPLITS_ARRIVED);
+        readiness.segments.emit(SDK_SEGMENTS_ARRIVED);
         return true;
       });
     } else {
